Move usePanelsMap from hooks/ to layout/
usePanelsMap is layout infrastructure that wires all panel components
into the layout system. It imports from every panel and belongs
alongside Layout.tsx, not in the generic hooks/ directory. Leave a
re-export at the old location for compatibility.
