+++ src/renderer/App.tsx
@@ -27,7 +27,7 @@ import { useFileNavigation } from './panels/fileViewer/hooks/useFileNavigation'
 import { useSessionLifecycle } from './features/sessions/hooks/useSessionLifecycle'
 import { useAgentChatStore } from './store/agentChat'
 import { useAppCallbacks } from './shared/hooks/useAppCallbacks'
-import { usePanelsMap } from './hooks/usePanelsMap'
+import { usePanelsMap } from './layout/usePanelsMap'
 import { useHelpMenu } from './shared/hooks/useHelpMenu'
 import { useSessionKeyboardCallbacks } from './shared/hooks/useSessionKeyboardCallbacks'
 import { focusSearchInput } from './shared/utils/focusHelpers'
